move tabs to templates folder
